split server and bot configs
