+++ env.bot.config.yml
@@ -32,31 +32,4 @@ SMF:
     FAUCET_IGNORE_LIST:
       description: "A list of Matrix accounts that will be silently (but logged) ignored, comma separated. Example: \"@alice:matrix.org,@bob:domain.com\""
       default: ""
-  BACKEND:
-    FAUCET_ACCOUNT_MNEMONIC:
-      description: "mnemonic seed from faucet account (create via polkadot.js.org)"
-      mandatory: true
-      masked: true
-      type: string
-    FAUCET_BALANCE_CAP:
-      description: "Upper limit cap on whether or not the account can receive more tokens. Defaults to 100."
-      default: 100
-      type: number
-    INJECTED_TYPES:
-      description: "if any type must be overridden"
-      default: {}
-      type: string
-    NETWORK_DECIMALS:
-      description: "decimal amount for the network"
-      default: 12
-      type: number
-    PORT:
-      description: "the port you want the server to listen on (should be part of the bot's `SMF_BOT_BACKEND_URL`.)"
-      default: 5555
-      type: number
-    RPC_ENDPOINT:
-      description: "ws rpc node endpoint"
-      default: "https://westend-rpc.polkadot.io/"
-      type: string
-
