+++ env.server.config.yml
@@ -1,0 +1,29 @@
+SMF:
+  BACKEND:
+    FAUCET_ACCOUNT_MNEMONIC:
+      description: "mnemonic seed from faucet account (create via polkadot.js.org)"
+      mandatory: true
+      masked: true
+      type: string
+    FAUCET_BALANCE_CAP:
+      description: "Upper limit cap on whether or not the account can receive more tokens. Defaults to 100."
+      default: 100
+      type: number
+    INJECTED_TYPES:
+      description: "if any type must be overridden"
+      default: {}
+      type: string
+    NETWORK_DECIMALS:
+      description: "decimal amount for the network"
+      default: 12
+      type: number
+    PORT:
+      description: "the port you want the server to listen on (should be part of the bot's `SMF_BOT_BACKEND_URL`.)"
+      default: 5555
+      type: number
+    RPC_ENDPOINT:
+      description: "ws rpc node endpoint"
+      default: "https://westend-rpc.polkadot.io/"
+      type: string
+
+
