+++ src/server/config.ts
@@ -1,0 +1,3 @@
+import { faucetConfig } from '../faucetConfig';
+
+export const config = faucetConfig('server');
